Move fields into a separate module. (#12)
* separate fields

* encapsulate constants

* update CHANGELOG

---------
